Minor: Improving code's coverage

+++ src/app/user/login/login.component.ts
@@ -3,6 +3,7 @@ import { ActivatedRoute, Router } from '@angular/router';
 import { User } from '../../models/index';
 import { UserService, AuthenticationService, AlertService } from '../../services/index';
 import { FormsModule } from '@angular/forms';
+import { TextMaskModule } from 'angular2-text-mask';
 
 @Component({
   selector: 'app-login',
@@ -38,22 +39,28 @@ export class LoginComponent implements OnInit {
 
     login() {
       this.loading = true;
-      this.authenticationService.login(this.email, this.password)
-          .subscribe(
-              result => {
-                  this.token = result;
-              },
-              error => {
-                  this.alertService.error(error);
-                  this.loading = false;
-                  if (error = 401) {
-                    console.log('Email ou senha inválidos');
-                    if (confirm('Email ou senha inválidos')) {}
-                  }
-              });
-          if (this.token) {
-            localStorage.setItem('token', this.token);
-            // this.router.navigate(['usuarios/editar/' + this.user.cod ]);
-          }
+      const emailMatch = this.email
+        .match(/^(([^<>()\[\]\\.,;:\s@"]+(\.[^<>()\[\]\\.,;:\s@"]+)*)|(".+"))@((\[[0-9]{1,3}\.[0-9]{1,3}\.[0-9]{1,3}\.[0-9]{1,3}])|(([a-zA-Z\-0-9]+\.)+[a-zA-Z]{2,}))$/);
+      if (emailMatch) {
+        this.authenticationService.login(this.email, this.password)
+        .subscribe(
+            result => {
+                this.token = result;
+            },
+            error => {
+                this.alertService.error(error);
+                this.loading = false;
+                if (error = 401) {
+                  console.log('Email ou senha inválidos');
+                  if (confirm('Email ou senha inválidos')) {}
+                }
+            });
+        if (this.token) {
+          localStorage.setItem('token', this.token);
+          // this.router.navigate(['usuarios/editar/' + this.user.cod ]);
+        }
+      } else {
+        console.warn('Não fez login');
+      }
   }
 }
